7장 리액티브 마이크로서비스 테스트 작성 테스트 Failed

+++ microservices/recommendation-service/src/main/java/se/magnus/microservices/core/recommendation/services/RecommendationServiceImpl.java
@@ -1,9 +1,12 @@
 package se.magnus.microservices.core.recommendation.services;
 
-import java.util.List;
+import static java.util.logging.Level.FINE;
+
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.web.bind.annotation.RestController;
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
 import se.magnus.api.core.recommendation.Recommendation;
 import se.magnus.api.core.recommendation.RecommendationService;
 import se.magnus.api.exceptions.InvalidInputException;
@@ -19,43 +22,59 @@ public class RecommendationServiceImpl implements RecommendationService {
     private final RecommendationRepository repository;
     private final RecommendationMapper mapper;
 
-    public RecommendationServiceImpl(ServiceUtil serviceUtil, RecommendationRepository repository, RecommendationMapper mapper) {
+    public RecommendationServiceImpl(ServiceUtil serviceUtil, RecommendationRepository repository,
+        RecommendationMapper mapper) {
         this.serviceUtil = serviceUtil;
         this.repository = repository;
         this.mapper = mapper;
     }
 
     @Override
-    public List<Recommendation> getRecommendations(int productId) {
+    public Flux<Recommendation> getRecommendations(int productId) {
 
         if (productId < 1) {
             throw new InvalidInputException("Invalid productId: " + productId);
         }
 
-        List<RecommendationEntity> entityList = repository.findByProductId(productId);
-        List<Recommendation> list = mapper.entityListToApiList(entityList);
-        list.forEach(e -> e.setServiceAddress(serviceUtil.getServiceAddress()));
+        log.info("Will get recommendations for product with id={}", productId);
 
-        log.debug("getRecommendations: response size: {}", list.size());
-        return list;
+        return repository.findByProductId(productId).log(log.getName(), FINE)
+            .map(mapper::entityToApi)
+            .map(this::setServiceAddress);
+    }
+
+    private Recommendation setServiceAddress(Recommendation e) {
+        e.setServiceAddress(serviceUtil.getServiceAddress());
+        return e;
     }
 
     @Override
-    public Recommendation createRecommendation(Recommendation body) {
-        try {
-            RecommendationEntity entity = mapper.apiToEntity(body);
-            RecommendationEntity newEntity = repository.save(entity);
+    public Mono<Recommendation> createRecommendation(Recommendation body) {
 
-            log.debug("createRecommendation: created a recommendation entity: {}/{}", body.getProductId(), body.getRecommendationId());
-            return mapper.entityToApi(newEntity);
-        } catch (DuplicateKeyException dke) {
-            throw new InvalidInputException("Duplicate key, Product Id: " + body.getProductId() + ", Recommendation Id:" + body.getRecommendationId());
+        if (body.getProductId() < 1) {
+            throw new InvalidInputException("Invalid productId: " + body.getProductId());
         }
+
+        RecommendationEntity entity = mapper.apiToEntity(body);
+        Mono<Recommendation> newEntity = repository.save(entity)
+            .log(log.getName(), FINE)
+            .onErrorMap(DuplicateKeyException.class,
+                ex -> new InvalidInputException(
+                    "Duplicate key, Product Id: " + body.getProductId()))
+            .map(e -> mapper.entityToApi(e));
+
+        return newEntity;
     }
 
     @Override
-    public void deleteRecommendations(int productId) {
-        log.debug("deleteRecommendations: tries to delete recommendations for the product with productId: {}", productId);
-        repository.deleteAll(repository.findByProductId(productId));
+    public Mono<Void> deleteRecommendations(int productId) {
+        if (productId < 1) {
+            throw new InvalidInputException("Invalid productId: " + productId);
+        }
+
+        log.debug(
+            "deleteRecommendations: tries to delete recommendations for the product with productId: {}",
+            productId);
+        return repository.deleteAll(repository.findByProductId(productId));
     }
 }
